requirements.txt
playwright
celery
django-axes
python-dotenv
requests
djangorestframework
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.27
sqlparse==0.5.5
typing-extensions==4.13.2
tzdata==2025.3
psycopg2-binary==2.9.7

